# docs/swagger.yaml
basePath: /
definitions:
...
  httptransport.GenerateInvoiceBody:
    properties:
      duplicate:
        type: boolean
      invoice:
        $ref: '#/definitions/httptransport.InvoiceBody'
      lines:
        items:
          $ref: '#/definitions/httptransport.InvoiceLineBody'
        type: array
    type: object
  httptransport.InvoiceBody:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      discount_minor:
        type: integer
      due_at:
        type: string
      finalized_at:
        type: string
      id:
        type: string
      number:
        type: string
      paid_at:
        type: string
      paid_minor:
        type: integer
      period_end:
        type: string
      period_start:
        type: string
      refunded_minor:
        type: integer
      status:
        type: string
      subscription_id:
        type: string
      subtotal_minor:
        type: integer
      tax_minor:
        type: integer
      tenant_id:
        type: string
      total_minor:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
  httptransport.InvoiceDetailBody:
    properties:
      invoice:
        $ref: '#/definitions/httptransport.InvoiceBody'
      lines:
        items:
          $ref: '#/definitions/httptransport.InvoiceLineBody'
        type: array
    type: object
...
  httptransport.InvoicePageBody:
    properties:
      items:
        items:
          $ref: '#/definitions/httptransport.InvoiceBody'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  httptransport.finalizeInvoiceRequest:
    properties:
      application_id:
        type: string
      due_at:
        type: string
      id:
        type: string
      tenant_id:
        type: string
      version:
        type: integer
    required:
    - application_id
    type: object
...
  httptransport.idVersionRequest:
    properties:
      id:
        type: string
      tenant_id:
        type: string
      version:
        type: integer
    type: object
  httptransport.invoicePeriodRequest:
    properties:
      application_id:
        type: string
      idempotency_key:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      subscription_id:
        type: string
      tenant_id:
        type: string
    required:
    - application_id
    type: object
...
info:
  contact: {}
  description: 'Plans, tenant subscriptions, usage-rated invoices, payments, and refunds.
    All business endpoints use POST+JSON and return code/message/body/request_id.
    Error ranges: 10000 common/input; 20000 authentication/authorization; 30000 business
    concurrency/idempotency; 50000 dependency/infrastructure.'
  title: Platform Billing Service API
  version: 0.1.0
paths:
  /api/v1/invoices/finalize:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice and due date
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.finalizeInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.InvoiceBody'
              type: object
      security:
      - Bearer: []
      summary: Finalize a draft invoice
      tags:
      - invoices
  /api/v1/invoices/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice period and idempotency key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.invoicePeriodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.GenerateInvoiceBody'
              type: object
      security:
      - Bearer: []
      summary: Generate an idempotent invoice
      tags:
      - invoices
  /api/v1/invoices/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.idVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.InvoiceDetailBody'
              type: object
      security:
      - Bearer: []
      summary: Get an invoice and line items
      tags:
      - invoices
  /api/v1/invoices/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.listInvoicesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.InvoicePageBody'
              type: object
      security:
      - Bearer: []
      summary: Search tenant invoices
      tags:
      - invoices
  /api/v1/invoices/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice period
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.invoicePeriodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.InvoiceDetailBody'
              type: object
      security:
      - Bearer: []
      summary: Preview an invoice using metering totals
      tags:
      - invoices
  /api/v1/invoices/void:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice and reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.voidInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.InvoiceBody'
              type: object
      security:
      - Bearer: []
      summary: Void an unpaid invoice
      tags:
      - invoices
...
  /api/v1/payments/apply-result:
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider result
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.applyPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ApplyPaymentResultBody'
              type: object
      security:
      - PSK: []
      summary: Apply an idempotent payment-provider result
      tags:
      - payments
  /api/v1/payments/create-attempt:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.createPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.CreatePaymentAttemptBody'
              type: object
      security:
      - Bearer: []
      summary: Create an idempotent payment attempt
      tags:
      - payments
  /api/v1/payments/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.getPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PaymentAttemptBody'
              type: object
      security:
      - Bearer: []
      summary: Get an application payment attempt
      tags:
      - payments
  /api/v1/payments/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.listPaymentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PaymentAttemptPageBody'
              type: object
      security:
      - Bearer: []
      summary: List application payment attempts
      tags:
      - payments
  /api/v1/payments/refunds/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.listPaymentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.RefundPageBody'
              type: object
      security:
      - Bearer: []
      summary: List application refunds
      tags:
      - payments
  /api/v1/payments/refunds/record:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refund
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.recordRefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.RecordRefundBody'
              type: object
      security:
      - Bearer: []
      summary: Record an idempotent refund
      tags:
      - payments
  /api/v1/plans/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.createPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PlanBody'
              type: object
      security:
      - Bearer: []
      summary: Create a billing plan
      tags:
      - plans
  /api/v1/plans/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plan identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.getPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PlanDetailBody'
              type: object
      security:
      - Bearer: []
      summary: Get a plan and its usage prices
      tags:
      - plans
  /api/v1/plans/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.listPlansRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PlanPageBody'
              type: object
      security:
      - Bearer: []
      summary: Search billing plans
      tags:
      - plans
  /api/v1/plans/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plan and current version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.updatePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.PlanBody'
              type: object
      security:
      - Bearer: []
      summary: Update a billing plan with optimistic locking
      tags:
      - plans
  /api/v1/plans/usage-prices/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Identity and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.idVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Delete a usage price with optimistic locking
      tags:
      - plans
  /api/v1/plans/usage-prices/upsert:
    post:
      consumes:
      - application/json
      parameters:
      - description: Usage price
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.upsertUsagePriceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.UsagePriceBody'
              type: object
      security:
      - Bearer: []
      summary: Create or update a metered usage price
      tags:
      - plans
  /api/v1/subscriptions/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Cancellation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.cancelSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SubscriptionBody'
              type: object
      security:
      - Bearer: []
      summary: Cancel a subscription immediately or at period end
      tags:
      - subscriptions
  /api/v1/subscriptions/change:
    post:
      consumes:
      - application/json
      parameters:
      - description: Plan change
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.changeSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SubscriptionBody'
              type: object
      security:
      - Bearer: []
      summary: Change a subscription immediately or next period
      tags:
      - subscriptions
  /api/v1/subscriptions/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.createSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SubscriptionBody'
              type: object
      security:
      - Bearer: []
      summary: Subscribe a tenant to a plan
      tags:
      - subscriptions
  /api/v1/subscriptions/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.getSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SubscriptionDetailBody'
              type: object
      security:
      - Bearer: []
      summary: Get a tenant subscription and plan
      tags:
      - subscriptions
  /api/v1/subscriptions/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.listSubscriptionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SubscriptionPageBody'
              type: object
      security:
      - Bearer: []
      summary: List tenant subscriptions
      tags:
      - subscriptions
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"